[refactor] reducerの分割
