+++ src/features/filters/action.ts
@@ -1,0 +1,8 @@
+export const CHANGE_FILTER = "filters/statusFilterChanged"
+
+type ChangeFilterAction = {
+  type: typeof CHANGE_FILTER;
+  payload: string;
+}
+
+export type FilterAction = ChangeFilterAction
