+++ src/features/filters/filtersSlice.ts
@@ -1,0 +1,22 @@
+import { CHANGE_FILTER, FilterAction } from "./action.ts"
+import { TodoFilter } from "./type.ts"
+
+const initialState: TodoFilter = {
+  status: "All",
+  colors: []
+}
+
+
+export const appReducer = (state = initialState, action: FilterAction) => {
+  switch (action.type) {
+    case CHANGE_FILTER: {
+      return {
+        ...state,
+        status: action.payload
+      }
+
+    }
+    default:
+      return state
+  }
+}
