+++ src/features/filters/type.ts
@@ -1,0 +1,4 @@
+export type TodoFilter = {
+  status: "All" | "Active" | "Completed"
+  colors: string[]
+}
