+++ src/features/todos/action.ts
@@ -1,6 +1,5 @@
 export const ADD_TODO = "todos/todoAdded"
 export const TOGGLE_TODO = "todos/todoToggled"
-export const CHANGE_FILTER = "filters/statusFilterChanged"
 
 type AddTodoAction = {
   type: typeof ADD_TODO;
@@ -12,9 +11,5 @@ type ToggleTodoAction = {
   payload: number;
 }
 
-type ChangeFilterAction = {
-  type: typeof CHANGE_FILTER;
-  payload: string;
-}
 
-export type TodoAction = AddTodoAction | ToggleTodoAction | ChangeFilterAction
+export type TodoAction = AddTodoAction | ToggleTodoAction
