+++ src/features/todos/type.ts
@@ -1,0 +1,6 @@
+export type Todo = {
+  id: number
+  text: string
+  completed: boolean
+  color?: string
+}
